+++ src/protean/server/observatory/routes/__init__.py
@@ -16,6 +16,7 @@ from .handlers import create_handlers_router
 from .infrastructure import create_infrastructure_router
 from .pages import create_page_router
 from .processes import create_processes_router
+from .timeline import create_timeline_router
 
 
 def create_all_routes(
@@ -36,5 +37,6 @@ def create_all_routes(
     api_router.include_router(create_processes_router(domains))
     api_router.include_router(create_eventstore_router(domains))
     api_router.include_router(create_infrastructure_router(domains))
+    api_router.include_router(create_timeline_router(domains))
 
     return page_router, api_router